Beginning of part 2
Implementation of twto menus.
Implementation of general functions of, change gray scale, change resolution and equalize a image.
Implementation of a function to describe image according to certain parameters.
Implementation of a function to create a sample file that will be used by the neural network.
Inicial implementation of a function to read that file and reorganize the data.